requirements.txt
streamlit>=1.35.0
requests>=2.31.0
pypdf>=4.2.0
pandas>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0
pytest>=8.0.0